backend/requirements.txt
matplotlib
yacs
numpy
django-rest-framework
django-cors-headers
django
pika
Pillow
opencv-python
PyGithub